Adding my changes to for the custom List View.

+++ src/searchFeature/nutritionAPIV2_model/Serving.java
@@ -7,7 +7,7 @@ import com.google.gson.annotations.SerializedName;
 public class Serving {
 
 	@SerializedName("qty")
-	public int quantity;
+	public float quantity;
 	
 	@SerializedName("uom")
 	public String uom;
